+++ src/db.rkt
@@ -1,51 +1,108 @@
 #lang racket
-;; database システムを作ろう。これと spin で、td-app を作る。
-;; td-app は今日やったこと、今日やることをネットから入力、検索できるアプリとする。
-;; ブラウザだけじゃなく、端末からも操作できるように。
-;; td は todo とも、todays ともどっちでも。
+;; database システムを作ろう。これと spin で、todo アプリを作る。
+;; todo アプリはやったこと、やることをネットから入力、検索できるアプリとする。
 ;;
 ;; * database ::= document の集まり
 ;; * document ::= entry の集まり
 ;; * entry    ::= (key entry) でキー、(valye entry) でバリューが取り出せるオブジェクト
 ;;
-;; require
-;; $ raco pkg install date
 
 (provide
-  init
-  insert
-  find
-  documents
-  load
-  save)
+  init       ; initialize database
+  documents  ; return whole data
+  insert     ; (insert 'key value 'key2 value2 ...)
+  find       ; (find comp-function 'key value)
+  has-key    ; (has-key 'key)
+  load       ; load saved documents into memory
+  save       ; save documents into file
+  now        ; returns iso-8601 timestamp
+  today      ; returns todays date string yyyy-mm-dd
+  )
 
-(require date)
+(require racket/date)
+(date-display-format 'iso-8601)
 
-(define nil '())
-(define *db* nil)
-
-(define init
-  (lambda ()
-    (set! *db* nil)))
-
-(init)
-
-(define key (lambda (entry) (car entry)))
-(define val (lambda (entry) (cdr entry)))
-
-(define make-entry
-  (lambda (key value)
-    (cons key value)))
-
+;; FIXME: function name?
 (define null-or-first
   (lambda (obj)
     (if (null? obj)
       '()
       (first obj))))
 
+(define now
+  (lambda () (date->string (current-date) #t)))
+
+(define today
+  (lambda () (substring (now) 0 10)))
+
+; (today)
+
+(define *db* '())
+
+(define documents (lambda () *db*))
+
+
+
+;; db はファイルに保存
+(define db-dat (string-append (path->string (current-directory)) "/db.dat"))
+
+(define init
+  (lambda ()
+    (set! *db* '())))
+
+;; id の定義は最初の load のあとで
+;; (define id (id-clojure)) のように。
+;; load に組み込むのが自然か？
+(define id-closure
+  (lambda ()
+    (let ((n (length (documents))))
+      (lambda ()
+        (set! n (+ n 1))
+        n))))
+
+(define save-to
+  (lambda (filename)
+    (call-with-output-file
+      filename
+      (lambda (out) (pretty-write *db* out))
+      #:exists 'replace)))
+
+(define save
+  (lambda () (save-to db-dat)))
+
+(define load-from
+  (lambda (filename)
+    (if (file-exists? filename)
+      (call-with-input-file
+        filename
+        (lambda (in) (set! *db* (read in))))
+      (set! *db* '()))))
+; serial number
+(define id #f)
+
+(define load
+  (lambda ()
+   (load-from db-dat)
+   (set! id (id-closure))))
+
+;;
+;; define database, documents, entries
+;;
+
+;; entries
+
+(define key first)
+(define val second)
+
+(define make-entry
+  (lambda (key value) (cons key (cons value '()))))
+
 (define find-entry
   (lambda (k doc)
     (null-or-first (filter (lambda (entry) (eq? k (key entry))) doc))))
+
+
+;; document には作成した日付 datetime と通し番号 id をつける。
 
 (define make-doc
   (lambda (key value . more)
@@ -55,81 +112,76 @@
         (make-entry key value)
         (apply make-doc more)))))
 
-(make-doc 'a 1)
-(make-doc 'a 1 'b 2)
-(make-doc 'a 1 'b 2 'c '(x y z))
-
-(find-entry 'b '((a . 1) (b . 2) (c x y z))) ;=> '(b . 2)
-(find-entry 'z '((a . 1) (b . 2) (c x y z))) ;=> '()
-
 (define make-datetime
- (lambda ()
-  (make-kv 'datetime (current-date-string-iso-8601 #t))))
-
-; (make-datetime)
+  (lambda () (make-entry 'datetime (date->string (current-date) #t))))
 
 (define add-datetime
-  (lambda (data)
-    (cons (make-datetime) data)))
+  (lambda (doc) (cons (make-datetime) doc)))
 
-; (add-datetime '((a . 1) (b . 2)))
+(define make-id
+  (lambda () (make-entry 'id (id))))
+
+(define add-id
+  (lambda (doc) (cons (make-id) doc)))
 
 (define insert
-  (lambda (data)
-    (set! *db* (cons (add-datetime data) *db*))))
+  (lambda (a . b)
+    (let* ((data (cons a b))
+           (doc (apply make-doc data)))
+      (set! *db* (cons (add-id (add-datetime doc)) *db*))
+      (save))))
 
-(insert (make-doc 'given-name "akari" 'family-name "kimura"))
-(insert (make-doc 'given-name "isana" 'family-name "kimura"))
-(insert (make-doc 'given-name "aoi" 'family-name "kimura"))
-(insert (make-doc 'given-name "hiroshi" 'family-name "kimura"))
-(insert (make-doc 'given-name "miyuki" 'family-name "kimura"))
-
+;; FIXME: funtion name?
 (define entry
   (lambda (k d)
     (null-or-first (filter (lambda (d) (eq? k (car d))) d))))
 
-(first *db*)
-(entry 'given-name (first *db*))
-
-(define find?
-  (lambda (doc f key value)
-    (let ((e (entry key doc)))
-      (and (not (empty? e)) (f value (val e))))))
-
-(find? (first *db*) string=? 'given-name "hiroshi")
-(find? (first *db*) string=? 'given-name "miyuki")
-
 (define find
-  (lambda (f key value)
-    (filter (lambda (doc) (find? doc f key value)) *db*)))
+  (lambda (fn key match)
+    (letrec
+      ((find-aux (lambda (doc fn key match)
+            (let ((e (entry key doc)))
+              (and (not (empty? e)) (fn (val e) match))))))
+        (filter (lambda (doc) (find-aux doc fn key match)) *db*))))
 
-(define documents (lambda () *db*))
+(define has-key
+  (lambda (key)
+    (find (lambda (k v) #t) key '())))
 
-(define save-to
-  (lambda (filename)
-    (call-with-output-file
-      filename
-      (lambda (out) (write *db* out))
-      #:exists 'replace)))
 
-(define save
-  (lambda () (save-to "scmdb.dat")))
+(define db-test
+  (lambda ()
+    ;; あとで戻せるように back に退避
+    (load)
+    (define back *db*)
+    (init)
+    ;;
+    (insert 'given-name "akari" 'family-name "kimura")
+    (insert 'given-name "isana" 'family-name "kimura")
+    (insert 'given-name "aoi" 'family-name "kimura")
+    (insert 'given-name "hiroshi" 'family-name "kimura")
+    (insert 'given-name "miyuki" 'family-name "kimura")
+    (insert 'wbc "japan" 'result "gold")
+    (insert 'wbc "usa" 'result "silver")
+    (insert 'computer 'mac 'os "macos")
+    (insert 'type "number" 'one 1)
+    (insert 'type "number" 'two 2)
+    ;;
+    (define out (open-output-string))
+    (display (find string=? 'given-name "aoi") out)
+    (display (first (find string=? 'family-name "kimura")) out)
+    (display (find string=? 'given-name "hiroshi") out)
+    (display (find string<? 'datetime "2023") out)
+    (display (find string<? 'datetime "2024") out)
+    (display (find string=? 'wbc "japan") out)
+    (display (has-key 'wbc) out)
+    (display (first (find string=? 'result "gold")) out)
+    (display (find = 'id 8) out)
+    ;;
+    (display (get-output-string out))
+    ;;
+    ;; 退避したデータベースを戻す
+    (set! *db* back)
+    ))
 
-(define load-from
-  (lambda (filename)
-    (call-with-input-file 
-      filename
-      (lambda (in) (set! *db* (read in))))))
-
-(define load
-  (lambda () (load-from "scmdb.dat")))
-
-(find string=? 'given-name "aoi")
-(first (find string=? 'family-name "kimura"))
-(find string=? 'given-name "hiroshi")
-(find string<? 'datetime "2023")
-(find string<? 'datetime "2024")
-
-(insert (make-doc 'wbc "japan" 'result "gold"))
-(find string=? 'wbc "japan")
-(save)
+; (db-test)